feat(components) move some tags to Head component
