+++ pages/case/[slug].js
@@ -6,7 +6,7 @@ import setComponentCounter from '../../components/_helpers/setParallaxComponentC
 import getDateFormat from '../../components/_helpers/getDateFormat';
 import scrollToElement from '../../components/_helpers/scrollToElement';
 
-import Head from '../../components/_helpers/head';
+import Head from '../../components/head/head';
 import PageHeader from '../../components/page-header/page-header';
 import MenuBar from '../../components/menu-bar/menu-bar';
 import TextCenter from '../../components/text-center/text-center';
