+++ pages/team/people.js
@@ -3,7 +3,7 @@ import '../../styles/index.less';
 import fetchContent from '../../lib/fetch-content';
 import withCacheControl from '../../lib/with-cache-control';
 
-import Head from '../../components/_helpers/head';
+import Head from '../../components/head/head';
 import MenuBar from '../../components/menu-bar/menu-bar';
 import PageHeader from '../../components/page-header/page-header';
 import TeamMembersOverview from '../../components/team-members-overview/team-members-overview';
